Fix bugs and change build system

+++ hal/generate_usage_reporting.py
@@ -3,7 +3,7 @@ import os
 
 def main():
     os.makedirs("hal/src/generated/main/native/include", exist_ok=True)
-    os.makedirs("hal/src/generated/main/native/java", exist_ok=True)
+    os.makedirs("hal/src/generated/main/java", exist_ok=True)
     usage_reporting_types_cpp = ""
     usage_reporting_instances_cpp = ""
     usage_reporting_types = ""
@@ -21,16 +21,23 @@ def main():
     with open("hal/src/generate/FRCNetComm.java.in") as java_usage_reporting:
         contents = java_usage_reporting.read()
         contents = contents.replace(r"${usage_reporting_types}", usage_reporting_types)
-        with open("hal/src/generated/main/native/java/FRCNetComm.java", "w") as java_out:
-            java_out.write(contents.replace(r"${usage_reporting_instances}", usage_reporting_instances))
+        if os.path.exists("hal/src/generated/main/java/FRCNetComm.java"):
+            with open("hal/src/generated/main/java/FRCNetComm.java", "w") as java_out:
+                java_out.write(contents.replace(r"${usage_reporting_instances}", usage_reporting_instances))
+        else:
+            with open("hal/src/generated/main/java/FRCNetComm.java", "x") as java_out:
+                java_out.write(contents.replace(r"${usage_reporting_instances}", usage_reporting_instances))
 
     with open("hal/src/generate/FRCUsageReporting.h.in") as cpp_usage_reporting:
         contents = cpp_usage_reporting.read()
-        contents = contents.replace(r"${usage_reporting_types_cpp}",
-                         usage_reporting_types_cpp)
-        with open("hal/src/generated/main/native/include/FRCUsageReporting.h", "w") as cpp_out:
-            cpp_out.write(contents.replace(
-                r"${usage_reporting_instances_cpp}", usage_reporting_instances_cpp))
+        contents = contents.replace(r"${usage_reporting_types_cpp}",usage_reporting_types_cpp).replace(
+            r"${usage_reporting_instances_cpp}", usage_reporting_instances_cpp)
+        if os.path.exists("hal/src/generated/main/native/include/FRCUsageReporting.h"):
+            with open("hal/src/generated/main/native/include/FRCUsageReporting.h", "w") as cpp_out:
+                cpp_out.write(contents)
+        else:
+            with open("hal/src/generated/main/native/include/FRCUsageReporting.h", "x") as cpp_out:
+                cpp_out.write(contents)
 
 
 if __name__ == "__main__":
